Support both HTTP & HTTPS
Right now the downloader does not support a http mirror, this can
complicate things when you want to use a custom internal mirror.

HTTPS might be the best option to use, but for a internal server this
might not the optimal solution as it brings forward awhole other issue
of selfsigned certs and configuration.

This fixes #1371