Support global proxy env vars as backup
Right now I'm using an older version of npm (3.10.8), where I have global environment variables for `HTTP_PROXY` and `HTTPS_PROXY` and `no_proxy` and they work just fine installing other modules. This module in particular doesn't support it. It just so happens when I add in the `http-proxy`, `https-proxy` and `no-proxy` settings to the NPM config so that puppeteer can pick them up, my `no-proxy` is unrecognized (as global or npm config). So I am stuck. There's no harm in supporting the global proxy settings also as a fallback if the npm ones aren't set.