feat: elHandle:screenshot captures full element
Modifies ElementHande::screenshot to capture the full element, even if
it's larger than the current viewport, or it falls off screen. Uses
window.scrollTo instead of scrollIntoViewIfNeeded because it must be
as much inside the viewport as possible. Partially visibile isn't good
for this use case. Resets the viewport size and scroll position after
it's finished.

Fixes #1779