Proposal: flexible chromium revisions
Often we want to test changes on new chromium versions. At the moment we would have to switch from the default downloader to specifying our own executable and downloading that etc. 

Exposing this to an environment variable would make it easy to test against different chromium versions in CI environments.