fix(chromium): bypass service workers when request interception is on
Puppeteer currently relies on invariant that once request interception
is enabled in DevTools protocol, every `Network.requestWillBeSent` event
is accompanied with a corresponding `Network.requestIntercepted`.

However, if a request is fulfilled by ServiceWorker, protocol will not
emit the `Network.requestIntercepted` event, and the invariant breaks.
As a result, Puppeteer never dispatches a `request` event.

The proper solution would be connecting to all the service workers and
adding an instrumentation to designate that request was fulfilled from
worker.

The compromised solution is to bypass service workers for all network
requests - and that's what we do here.

Fix #4041