feat(Launcher): add 'detached' option to launch
This option causes launch to run the browser in a detached state and return null. You can use 'connect' to use the launched instance.

To connect to it you need to know the websocket endpoint so this is best used with the --remote-debugging-port option so that you can query the browser to get the endpoint.