chore: Upgrade to TypeScript 3.5
TS 3.5 got much stricter on writing changes to objects with varied types
[1] so we have to do a bit of typecasting work to convince TS about the
types of keys and values that we are setting.

Longer term we should think about a better data structure that avoids us
having to jump through some hoops but for now I think this is a
reasonable step to get us onto 3.5.

Same story regarding bindings on `window`, the easiest fix is to cast
window to any for the code that adds to it. I'm sure we can come up with
a more type-safe way of doing this in the future.

[1]: https://github.com/Microsoft/TypeScript/wiki/Breaking-Changes#fixes-to-unsound-writes-to-indexed-access-types