chore: Upgrade TypeScript to 3.6
Continues the work to get up to TS 3.8 (latest release at time of
writing).

This version of TS introduced built in definitions for web workers that
include an `interface Worker` so TS gets confused when it sees us
reference a `Worker`. I have renamed the imports to `PuppeteerWorker` as
I couldn't figure out a way to tell TS to not load in the worker types;
longer term we might consider renaming `Worker` to `PuppeteerWorker` (or
an alternative) but that would be a breaking change that we don't need
right now.

The other fix is similar; TypeScript doesn't differentiate between the
built-in `WebSocket` type and the `ws` library. Renaming the import
solves this too.