chore: Upgrade to TypeScript 3.7
TypeScript seems to struggle to understand `Promise.all` when the items
in the array return different types. If we were authoring in TS we could
fix this with TS generics (`Promise.all<OurTypeHere>(...)`) but for now
we can typecast the result. We'll fix this properly when we author in
TS.