chore: Fix Page.emulateMedia doclint failure
This is expected as we now alias `emulateMedia` in `index.js` which
isn't a file checked by DocLint. We alias there to avoid having the
function overriden by the `asyncInstallHooks` code.

This commit updates doclint to know about methods that we expect it will
find are missing and in that case just skip over them. We should only do
this for methods where we plan to deprecate them or we have to define
them in an odd way to work around some problem (and if that's the case
long term we should fix that problem so we can define them as normal).