feat(TypeScript): Move DeviceDescriptors to TS
This commit moves `src/DeviceDescriptors` to be authored in TypeScript.
This file was chosen due to its simplicity so that we can focus on
getting a mixed JS/TS codebase playing nicely before migrating the more
complex files.

The file itself was a bit odd: although the array of devices was
exported via `module.exports` that was never referenced by any
consumers; each device was also exported via `module.exports[name] =
device` and that is how it's consumed. The Puppeteer docs suggest using
it like so:

```js
puppeteer.devices['iPhone 6']
```

So instead of exporting the array and then setting a bunch of properties
on that, we instead define the array and export an object of keys where
each key is a device. This is a breaking change (see the footer for
details).

Rather than export an object I'd much rather export a Map, but that
would be a larger breaking change and I'm keen to avoid those for the time
being.

Note that we have to use special TypeScript specific syntax for the
export that enables it to work in a CommonJS codebase [1] but again I'd
rather this than move to ESM at this time. TypeScript still outputs
CommonJS into `lib/` as you would expect.

BREAKING CHANGE: We no longer export an array of devices, so any users
relying on doing:

```js
puppeter.devices.forEach(...)
```

Will now see a breakage. The fix is to use
`Object.{keys/entries/values}` to iterate instead.

[1]: https://www.typescriptlang.org/docs/handbook/modules.html#export--and-import--require