chore: Migrate tests to Mocha
Rather than maintain our own test runner we should instead lean on the
community and use Mocha which is very popular and also our test runner
of choice in DevTools too.

Note that this commit doesn't remove the TestRunner source as it's still
used for other unit tests, but they will be updated in a future PR and
then we can remove the TestRunner.

The main bulk of this PR is updating the tests as the old TestRunner
passed in contextual data via the `it` function callback whereas Mocha
does not, so we introduce some helpers for the tests to make it easier.