chore: remove puppeteer-web
We don't support it and v3 shipped without including puppeteer-web in
the browser. People are welcome to manually use Browserify to try to get
Puppeteer running in a browser but it ultimately isn't our primary focus
right now.

Getting puppeteer-core able to run in a browser is something we'll be
looking at in the future so we'll revisit this soon.