chore: Error for bad Node versions
We define the Node version in the `engines` field in `package.json`
however this is only used and mentioned to the user when installing the
package as a dependency [1]. Node/npm is quite noisy when installing
these so it's very easily missed.

We now use the semver package to check our version in the install script
and bail if it's not at least what's in our `engines` field. This fix
isn't perfect because if we use syntax that isn't supported in old
versions the user will get a syntax error rather than our nice error,
but it should help moving forwards as we push Node versions up. e.g.
when we were to drop Node 10 (for the avoidance of doubt, not happening
anytime soon at all) this would help users be aware and know to upgrade.

[1]: https://docs.npmjs.com/files/package.json#engines