chore: migrate tests to TypeScript.
This CL migrates all the tests to TypeScript. The main benefits of this
is that we start consuming our TypeScript definitions and therefore find
errors in them. The act of migrating found some bugs in our definitions
and now we can be sure to avoid them going forwards.

You'll notice the addition of some `TODO`s in the code; I didn't want
this CL to get any bigger than it already is but I intend to follow
those up once this lands. It's mostly figuring out how to extend the
`expect` types with our `toBeGolden` helpers and some other slight
confusions with types that the tests exposed.